fix(report-card): mobile relocates vote buttons into the signal-strip row
Every .report-block has a .signal-strip inside .card-body and a .vote-btns
inside .card > .right. Because .card stacks vertically on mobile, the vote
row landed a full row below the icon strip with a visible gap.

A small relocator moves the vote-btns node into the signal-strip when the
viewport is <= 560px (and back into .right when it widens). Called from
render() so re-renders after sort / filter / vote reapply, and a debounced
window resize handler re-runs it both directions.

Closes #165
